remove bangle.io
