+++ .eslintignore
@@ -1,4 +1,3 @@
 # Third party code
 components/common/CharmEditor/components/@bangle.dev
-components/common/CharmEditor/components/@bangle.io
 components/common/BoardEditor/focalboard
