--- components/common/CharmEditor/components/@bangle.io/lib/config/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "@bangle.io/config",
-  "version": "0.0.0",
-  "authors": [
-    {
-      "name": "Kushan Joshi",
-      "email": "0o3ko0@gmail.com",
-      "web": "http://github.com/kepta"
-    }
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/bangle-io/bangle-io.git",
-    "directory": "lib/config"
-  },
-  "bugs": {
-    "url": "https://github.com/bangle-io/bangle-io/issues"
-  },
-  "main": "index.ts",
-  "module": "index.ts",
-  "publishConfig": {
-    "access": "public"
-  },
-  "dependencies": {
-    "@bangle.io/constants": "workspace:*",
-    "@bangle.io/shared-types": "workspace:*"
-  },
-  "license": "AGPL-3.0-or-later"
-}
