--- components/common/CharmEditor/components/@bangle.io/lib/ui-components/UniversalPalette/PaletteItemsContainer.tsx
@@ -1,14 +1,0 @@
-import React, { ReactNode } from 'react';
-
-export function PaletteItemsContainer({ children }: { children: ReactNode }) {
-  return (
-    <div
-      className="universal-palette-items-container"
-      style={{
-        overflowY: 'scroll',
-      }}
-    >
-      {children}
-    </div>
-  );
-}
