--- components/common/CharmEditor/components/@bangle.io/lib/ui-components/UniversalPalette/UniversalPalette.css
@@ -1,123 +1,0 @@
-:root {
-  --universal-palette-alignment-spacing: 0.5rem;
-  --universal-palette-item-disabled-bgColor: rgb(232, 232, 232);
-  --universal-palette-item-divider-width: 1px;
-  --universal-palette-item-disabled-bg-text-color: rgb(161, 161, 161);
-}
-
-.universal-palette-item .b-left-node {
-  display: inline-flex;
-  flex-direction: row;
-  justify-content: center;
-}
-.universal-palette-item .b-right-node {
-  display: inline-flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.universal-palette-item:hover .b-right-node {
-  display: none;
-}
-
-.universal-palette-item .b-right-hover-node {
-  display: none;
-}
-
-.universal-palette-item:hover .b-right-hover-node {
-  display: inline-flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.universal-palette-container.widescreen {
-  top: 0;
-}
-
-.universal-palette-container {
-  background-color: var(--universalPalette-bgColor);
-  margin-left: auto;
-  margin-right: auto;
-  max-width: var(--universalPalette-maxWidth);
-  width: 100%;
-  max-height: min(75vh, 400px);
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 0.25rem;
-  z-index: 30;
-  position: fixed;
-  left: 0;
-  right: 0;
-  top: var(--activitybar-width);
-  box-shadow: 0 0 rgba(26, 24, 24, 0), 0 0 #0000, 0 0 #0000, 0 0 #0000,
-    0 25px 50px -12px rgba(0, 0, 0, 0.25);
-}
-
-.universal-palette-input-wrapper {
-  padding: var(--universal-palette-alignment-spacing);
-  border-bottom: 1px solid var(--window-borderColor-0);
-  min-height: 3.2rem;
-}
-
-.universal-palette-input {
-  min-height: 28px;
-  padding: calc(var(--universal-palette-alignment-spacing) / 2)
-    var(--universal-palette-alignment-spacing);
-}
-
-.universal-palette-input-wrapper .b-left-node {
-  margin-right: calc(var(--universal-palette-alignment-spacing) * 1.5);
-  margin-left: var(--universal-palette-alignment-spacing);
-  display: flex;
-  align-items: center;
-}
-
-/* palette-items-container */
-
-/* single line item */
-.universal-palette-item {
-  padding: 0 calc(2 * var(--universal-palette-alignment-spacing));
-}
-
-.universal-palette-item.b-divider {
-  border-top: var(--universal-palette-item-divider-width) solid
-    var(--window-borderColor-0);
-}
-
-.universal-palette-item:hover {
-  background-color: var(--accent-secondary);
-}
-
-.universal-palette-item.active {
-  background-color: var(--accent-primary-1);
-}
-
-.universal-palette-item.disabled {
-  color: var(--textColor-1);
-  cursor: not-allowed;
-}
-
-.universal-palette-item.disabled.active {
-  background-color: var(--universal-palette-item-disabled-bgColor);
-}
-
-.universal-palette-item.disabled:hover {
-  background-color: var(--universal-palette-item-disabled-bgColor);
-}
-
-.universal-palette-item .b-left-node {
-  margin-right: var(--universal-palette-alignment-spacing);
-}
-
-.universal-palette-item .b-extra-info {
-  margin-left: var(--universal-palette-alignment-spacing);
-  color: var(--textColor-1);
-}
-
-.universal-palette-item .b-extra-info.extra-info-on-new-line {
-  margin-left: 0;
-}
-
-.universal-palette-item .b-description {
-  color: var(--textColor-1);
-}
