--- components/common/CharmEditor/components/@bangle.io/lib/ui-components/UniversalPalette/UniversalPalette.tsx
@@ -1,6 +1,0 @@
-export * from './hooks';
-export * from './PaletteContainer';
-export * from './PaletteInfo';
-export * from './PaletteInput';
-export * from './PaletteItem';
-export * from './PaletteItemsContainer';
