--- components/common/CharmEditor/components/@bangle.io/lib/ui-components/UniversalPalette/index.ts
@@ -1,1 +1,0 @@
-export * as UniversalPalette from './UniversalPalette';
