--- components/common/CharmEditor/components/@bangle.io/lib/ui-components/index.ts
@@ -1,1 +1,0 @@
-export * from "./UniversalPalette";
