--- components/common/CharmEditor/components/@bangle.io/lib/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './hooks';
-export * from './safe-js-callbacks';
-export * from './utility';
-
