+++ components/common/CharmEditor/components/nestedPage/components/NestedPagesList.tsx
@@ -46,9 +46,6 @@ export default function NestedPagesList () {
 
         return (
           <MenuItem
-            sx={{
-              background: theme.palette.background.light
-            }}
             onClick={() => onSelectPage(page)}
             key={page.id}
           >
