+++ components/common/CharmEditor/keybindings.ts
@@ -1,8 +1,8 @@
-import { isMac } from './is-mac';
 
+const isMac = typeof navigator !== 'undefined' ? /Mac/.test(navigator.platform) : false;
 const altInMac = '⌥'; // option
 
-export function keyDisplayValue(key: string) {
+export function keyDisplayValue (key: string) {
   if (key.includes('Mod')) {
     key = key.split('Mod').join(isMac ? '⌘' : 'Ctrl');
   }
@@ -25,26 +25,28 @@ export function keyDisplayValue(key: string) {
 
 class KeyBinding {
   key: string;
-  constructor({ key }: { key: string }) {
+
+  constructor ({ key }: { key: string }) {
     this.key = key;
   }
-  get displayValue() {
+
+  get displayValue () {
     return keyDisplayValue(this.key);
   }
 }
 
 export const keybindings = {
   toggleSecondaryEditor: new KeyBinding({
-    key: 'Mod-\\',
+    key: 'Mod-\\'
   }),
 
   toggleInlineCommandPalette: new KeyBinding({
-    key: 'Mod-/',
+    key: 'Mod-/'
   }),
   newNote: new KeyBinding({
-    key: 'Ctrl-n',
+    key: 'Ctrl-n'
   }),
   searchNotes: new KeyBinding({
-    key: 'Mod-F',
-  }),
+    key: 'Mod-F'
+  })
 };
